+++ test/api/UserManagement/user_management_test.go
@@ -1,37 +1,30 @@
 package api_test
 
 import (
-	"github.com/stretchr/testify/require"
 	"testing"
-//	"os"
-	"github.com/Telmate/proxmox-api-go/test/api"
+
+	"github.com/stretchr/testify/require"
+
 	pxapi "github.com/Telmate/proxmox-api-go/proxmox"
+	api_test "github.com/Telmate/proxmox-api-go/test/api"
 )
 
 var user = pxapi.ConfigUser{
 	User: pxapi.UserID{
 		Name:  "Bob",
 		Realm: "pve",
-		},
-		Comment:   "",
-		Email:     "bob@example.com",
-		Enable:    true,
-		FirstName: "Bob",
-		LastName:  "Bobson",
-}
-
-func Test_List_Users(t *testing.T) {
-	Test := api_test.Test{}
-	_ = Test.CreateTest()
-	users, err := pxapi.ListUsers(Test.GetClient(), false)
-	require.NoError(t, err)
-	require.Equal(t, 1, len(*users))
+	},
+	Comment:   "",
+	Email:     "bob@example.com",
+	Enable:    true,
+	FirstName: "Bob",
+	LastName:  "Bobson",
 }
 
 func Test_Create_User(t *testing.T) {
-	Test := api_test.Test {}
+	Test := api_test.Test{}
 	_ = Test.CreateTest()
-    err := user.CreateUser(Test.GetClient())
+	err := user.CreateUser(Test.GetClient())
 	require.NoError(t, err)
 }
 
@@ -51,7 +44,7 @@ func Test_User_Is_Added(t *testing.T) {
 }
 
 func Test_Remove_User(t *testing.T) {
-	Test := api_test.Test {}
+	Test := api_test.Test{}
 	_ = Test.CreateTest()
 	err := user.DeleteUser(Test.GetClient())
 	require.NoError(t, err)
